+++ src/pages/tournament/components/tournament/tournament.component.js
@@ -29,14 +29,32 @@ module.exports = {
                 acc[match.id] = match;
                 return acc;
             }, {}));
-        }        
+        }
+
+
+
         this.state = {
             hasStatistics: !_.isEmpty(input.statistics.best_three_dart_avg),
             matches: matches,
             matchesMap: matchesMap,
             unplayed: unplayed,
             groups: Array.from(groups),
-            overview: overview
+            overview: overview,
+            unq_statistics: {},
+            statistics: input.statistics,
+            showAllStats: true
+        }
+        if (this.state.hasStatistics) {
+            // Create unique statistics
+            if (input.statistics.best_301_darts_thrown) {
+                this.state.unq_statistics.best_301_darts_thrown  = _.uniq(input.statistics.best_301_darts_thrown, false, item => item.player_id);
+            }
+            if (input.statistics.best_501_darts_thrown) {
+                this.state.unq_statistics.best_501_darts_thrown  = _.uniq(input.statistics.best_501_darts_thrown, false, item => item.player_id);
+            }
+            this.state.unq_statistics.best_first_nine_avg    = _.uniq(input.statistics.best_first_nine_avg, false, item => item.player_id);
+            this.state.unq_statistics.best_three_dart_avg    = _.uniq(input.statistics.best_three_dart_avg, false, item => item.player_id);
+            this.state.unq_statistics.checkout_highest       = _.uniq(input.statistics.checkout_highest, false, item => item.player_id);
         }
     },
 
@@ -79,5 +97,11 @@ module.exports = {
     onUpdatePredictions(groupId, overview) {
         const comp = this.getComponent(`predictor-overview-${groupId}`);
         comp.updateStandings(overview);
+    },
+    onToggleStats(value, event) {
+        this.state.showAllStats = value;
+        this.state.statistics = value ? this.input.statistics : this.state.unq_statistics;
+        this.setStateDirty("statistics");
+        console.log(this.state.statistics);
     }
 }